Minor changes
Changes:
- add log message when reader closes in socketclient
- tweak colors in govpp diff command
- code cleanup in codec package
- enable debug profiling in tests
- fix integration test README
- detect VPP crashes in integration tests
